Move get_module_and_compile_dependencies to codegen.utilities
